Completing lesson 1.6 adding in event listeners and manipulating the DOM
Adding in styling, and basic html tags for the creation of my Pokémon app. Including functions to create elements and append them to the DOM.

Also adding an event listener to each button that will listen for a click and run a function that console.log()'s the Pokémon objects key value pairs.